Refactoring. Removed need for Role table. Renamed models after TCAT documentation.
