--- src/main/java/com/anothercaffeinatedday/rjug/model/UsersRoles.java
@@ -1,62 +1,0 @@
-/*
- * Copyright (C) 2012 Timothy Stone
- * 
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- * 
- *   http://www.apache.org/licenses/LICENSE-2.0
- * 
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License. * 
- */
-package com.anothercaffeinatedday.rjug.model;
-
-import java.io.Serializable;
-import javax.persistence.Entity;
-import javax.persistence.ManyToOne;
-
-/**
- *
- * @author tstone
- */
-
-@Entity
-public class UsersRoles extends ModelBase {
-    
-    @ManyToOne
-    private String username;
-    @ManyToOne
-    private String rolename;
-
-    /**
-     * @return the username
-     */
-    public String getUsername() {
-        return username;
-    }
-
-    /**
-     * @param username the username to set
-     */
-    public void setUsername(String username) {
-        this.username = username;
-    }
-
-    /**
-     * @return the rolename
-     */
-    public String getRolename() {
-        return rolename;
-    }
-
-    /**
-     * @param rolename the rolename to set
-     */
-    public void setRolename(String rolename) {
-        this.rolename = rolename;
-    }
-}
